+++ R/deploy_git-r6.R
@@ -4,7 +4,7 @@
 #' @export
 check_deploy_github = R6::R6Class(
   "check_deploy_github",
-  inherit = uatBase::base_check,
+  inherit = audit.base::base_check,
   public = list(
     #' @description Checks deployment of an R Markdown document with PDF output
     #' @param debug_level See check() for details