+++ R/deploy_plumber-r6.R
@@ -4,13 +4,13 @@
 #' @export
 check_deploy_plumber_api = R6::R6Class(
   "check_deploy_pluumber_api",
-  inherit = uatBase::base_check,
+  inherit = audit.base::base_check,
   public = list(
     #' @description Checks deployment of a Plumber API
     #' @param debug_level See check() for details
     check = function(debug_level) {
       plumber_dir = system.file("extdata", private$group,
-                                package = "jrHealthCheckConnect", mustWork = TRUE)
+                                package = "audit.connect", mustWork = TRUE)
       private$checker(deploy_plumber(plumber_dir, debug_level = debug_level))
       return(invisible(NULL))
     }