+++ R/deploy_streamlit-r6.R
@@ -4,7 +4,7 @@
 #' @export
 check_deploy_python_streamlit = R6::R6Class(
   "check_deploy_python_streamlit",
-  inherit = uatBase::base_check,
+  inherit = audit.base::base_check,
   public = list(
     #' @description Checks deployment of Python Streamlit app
     #' @param debug_level See check() for details
@@ -14,7 +14,7 @@ check_deploy_python_streamlit = R6::R6Class(
         return(invisible(NULL))
       }
       python_dir = system.file("extdata", private$group, private$short,
-                        package = "jrHealthCheckConnect", mustWork = TRUE)
+                        package = "audit.connect", mustWork = TRUE)
       private$checker(deploy_python(python_dir,
                                     python_files = "app.py",
                                     rsconnect_type = "streamlit",